Yusef update-
I worked on implenting the siphon feature. I also worked on the excell upload and changing the user roles by a scroll wheel instead of buttons.

front/src/pages/group.jsx

in here I implented the code for showing the buttons for the siphon feature. That would allow all students in the group be be able to see and vote on the siphon active request. I also added the scrool wheel code for the teacehr side to be able to change the user roles.

backend/models/Group.js

This is where I did the backend for the voting on the siphon request

backend/routes/group.js

this is where i implemented the code for the siphon request and the teacher overide. Which allows the teacher to have the final say in the matter.

back/utils/transferbits.js

I created this file to handle the transfering of the bits from the siphon feature.

frontend/src/components/siphonmodal.jsx

This is where the logic and frontend set up was made for the siphon request on the student side.